refactor: Update comments and variable names in get-products-to-db

+++ src/get-products-to-db.ts
@@ -4,17 +4,27 @@ import { compareArrAndReturnNew } from "../utils/compare";
 
 Promise.resolve(getNewProducts())
   .then(async (result) => {
+    // result 本身就是快照
     try {
       const { db } = await connectToDatabase();
       const temp = (await db.collection("temp").find({}).toArray()) || [];
-      if (temp.length <= 0) {
+      const activeStore =
+        (await db.collection("store").find({}).toArray()) || [];
+      if (temp.length <= 0 && activeStore.length <= 0) {
         // 如果 temp 原本就空的，那就不用比對，把這次的資料作為 init 存為 temp
         await db.collection("temp").drop();
         await db.collection("temp").insertMany(result);
         return;
       }
+      // 1. 比對和上一次的暫存是新商品
       // @ts-ignore TODO: fix type error
-      const newProducts = compareArrAndReturnNew(temp, result);
+      const productToCompareWithTemp = compareArrAndReturnNew(temp, result);
+      // 2. 就算是新商品，也要看 store 是否出現過
+      const newProducts = compareArrAndReturnNew(
+        // @ts-ignore TODO: fix type error
+        activeStore,
+        productToCompareWithTemp
+      );
       if (newProducts.length <= 0) {
         // 刪除原本的 temp，目的是作為清空來用
         await db.collection("temp").drop();
@@ -34,7 +44,7 @@ Promise.resolve(getNewProducts())
         .insertMany(newProductsWithState);
       // 刪除原本的 temp，目的是作為清空來用
       const dbTempDropRes = await db.collection("temp").drop();
-      // 把這次收集到的新品資料，放到新的 temp
+      // 把這次收集到的新品資料，放到新的 temp，這就是當天的快照
       const dbResponse = await db.collection("temp").insertMany(result);
     } catch (err) {
       console.log(err);
